+++ aggregator/src/test/scala/weco/concepts/aggregator/ConceptExtractorTest.scala
@@ -1,96 +1,13 @@
 package weco.concepts.aggregator
 import scala.io.Source
-
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.GivenWhenThen
 import org.scalatest.LoneElement.convertToCollectionLoneElementWrapper
 import org.scalatest.featurespec.AnyFeatureSpec
 import org.scalatest.prop.TableDrivenPropertyChecks
 import ujson.ParseException
-//import weco.concepts.aggregator.SourceConcept.{aCanonicalId, aLabel, aType, anAuthority, anExternalId}
-import weco.concepts.common.model._
+import weco.concepts.aggregator.testhelpers.SourceConcept
 
-import scala.util.Random
-
-trait ValueGenerators {
-  private val keys = IdentifierType.typeMap.keys.toList
-
-  def anAuthority: String =
-    keys(Random.nextInt(keys.length))
-
-  private val terms = List(
-    "Lorem",
-    "Ipsum",
-    "Dolor",
-    "Sit",
-    "Amet",
-    "Consectetur"
-  )
-
-  def aLabel: String =
-    terms(Random.nextInt(terms.length))
-
-  def aType: String =
-    ConceptExtractor.conceptTypes(
-      Random.nextInt(ConceptExtractor.conceptTypes.length)
-    )
-
-  def anExternalId: String =
-    "EXT_" + Random.alphanumeric.take(10).mkString
-
-  def aCanonicalId: String =
-    Random.alphanumeric.take(8).mkString
-}
-
-/*
- * Utility object to generate concepts in the form in which they are
- * presented in the Works API.  This is the form in which they arrive
- * into the ConceptExtractor
- */
-case class SourceConcept(
-  authority: String,
-  identifier: String,
-  label: String,
-  canonicalId: String,
-  ontologyType: String
-) {
-  override def toString: String =
-    s"""
-       |{
-       |  "id": "$canonicalId",
-       |  "identifiers": [
-       |    {
-       |      "identifierType": {
-       |        "id": "$authority",
-       |        "label": "This field is ignored",
-       |        "type": "IdentifierType"
-       |      },
-       |      "value": "$identifier",
-       |      "type": "Identifier"
-       |    }
-       |  ],
-       |  "label": "$label",
-       |  "type": "$ontologyType"
-       |}
-       |""".stripMargin
-}
-
-object SourceConcept extends ValueGenerators {
-  def apply(
-    authority: String = anAuthority,
-    identifier: String = anExternalId,
-    label: String = aLabel,
-    canonicalId: String = aCanonicalId,
-    ontologyType: String = aType
-  ): SourceConcept = new SourceConcept(
-    authority,
-    identifier,
-    label,
-    canonicalId,
-    ontologyType
-  )
-
-}
 
 class ConceptExtractorTest
     extends AnyFeatureSpec
@@ -105,8 +22,60 @@ class ConceptExtractorTest
     info("- the name used in the original document")
     info("- the wellcome canonical identifier")
 
-    // TODO: also all identifiertypes.
-    val t = Table(
+    Scenario("a real example") {
+      // Choose an actual document from the API output and
+      // prove that it works.
+      Given("a Work from the catalogue API")
+      val json = Source.fromResource("uk4kymkq.json").getLines().mkString("\n")
+      val concepts = ConceptExtractor(json)
+      // The more precisely defined tests here show what should be
+      // extracted.  This is just to demonstrate that it can successfully
+      // find the concepts in a real document from the catalogue api.
+      Then("all the concepts in the work are returned")
+      concepts.length shouldBe 6
+    }
+
+    Scenario(s"extract the data from a concept") {
+      val identifierType = "lc-subjects"
+      val identifier = "12345678900"
+      val label = "Quirkafleeg"
+      val canonicalId = "b7yui912"
+      val ontologyType = "Concept"
+      Given(s"a document containing a concept of type $ontologyType")
+      And(s"an identifierType of $identifierType")
+      And(s"an identifier of $identifier")
+      And(s"a label $label")
+      And(s"a canonical identifier $canonicalId")
+      val json =
+        s"""{
+           |"concepts":[
+            ${SourceConcept(
+          authority = identifierType,
+          identifier = identifier,
+          label = label,
+          canonicalId = canonicalId,
+          ontologyType = ontologyType
+        )}
+           |]
+           |}""".stripMargin
+
+      Then("there is one resulting concept")
+      val concept = ConceptExtractor(json).loneElement
+
+      And(s"the concept's identifierType is $identifierType")
+      concept.identifier.identifierType.id shouldBe identifierType
+
+      And(s"the concept's identifier is $identifier")
+      concept.identifier.value shouldBe identifier
+
+      And(s"the concept's canonicalIdentifier is $canonicalId")
+      concept.canonicalId shouldBe canonicalId
+
+      And(s"the concept's label is $label")
+      concept.label shouldBe label
+    }
+
+    val ontologyTypes = Table(
       "ontologyType",
       "Concept",
       "Person",
@@ -114,51 +83,62 @@ class ConceptExtractorTest
       "Meeting",
       "Period"
     )
-    forAll(t) { ontologyType =>
-      Scenario(s"Extracts the data from a concept of type $ontologyType ") {
-        val identifierType = "lc-subjects"
-        val identifier = "12345678900"
-        val label = "Quirkafleeg"
-        val canonicalId = "b7yui912"
+    forAll(ontologyTypes) { ontologyType =>
+      Scenario(s"extract a concept of type $ontologyType ") {
         Given(s"a document containing a concept of type $ontologyType")
-        And(s"an identifierType of $identifierType")
-        And(s"an identifier of $identifier")
-        And(s"a label $label")
-        And(s"a canonical identifier $canonicalId")
+        val sourceConcept = SourceConcept(
+          ontologyType = ontologyType
+        )
         val json =
           s"""{
              |"concepts":[
-            ${SourceConcept(
-              authority = identifierType,
-              identifier = identifier,
-              label = label,
-              canonicalId = canonicalId,
-              ontologyType = ontologyType
-            )}
+            $sourceConcept
              |]
              |}""".stripMargin
 
-        Then("there is one resulting concept")
-        val concept = ConceptExtractor(json).loneElement
-
-        And(s"the concept's identifierType is $identifierType")
-        concept.identifier.identifierType.id shouldBe identifierType
-
-        And(s"the concept's identifier is $identifier")
-        concept.identifier.value shouldBe identifier
-
-        And(s"the concept's canonicalIdentifier is $canonicalId")
-        concept.canonicalId shouldBe canonicalId
-
-        And(s"the concept's label is $label")
-        concept.label shouldBe label
+        Then("that concept is extracted")
+        ConceptExtractor(json).loneElement should have(
+          Symbol("label") (sourceConcept.label),
+          Symbol("canonicalId") (sourceConcept.canonicalId)
+        )
       }
+    }
+
+  val identifierTypes = Table(
+    "identifierType",
+    "lc-subjects",
+    "lc-names",
+    "nlm-mesh",
+    "label-derived",
+  )
+  forAll(identifierTypes) { identifierType =>
+    Scenario(s"extract a concept with an identifier in the $identifierType scheme ") {
+      Given(s"a document containing a concept with an identifier of type $identifierType")
+      val sourceConcept = SourceConcept(
+        authority = identifierType
+      )
+      val json =
+        s"""{
+           |"concepts":[
+            $sourceConcept
+           |]
+           |}""".stripMargin
+
+      Then("that concept is extracted")
+      val concept = ConceptExtractor(json).loneElement
+      concept should have(
+        Symbol("label") (sourceConcept.label),
+        Symbol("canonicalId") (sourceConcept.canonicalId),
+      )
+      concept.identifier.identifierType.id shouldBe identifierType
     }
   }
 
-  Scenario("Multiple different concepts throughout the document") {
+
+
+  Scenario("extract multiple different concepts throughout the document") {
     info("Concepts may be found in various places in a document")
-    info("The extractor is agnostic as to where they might be")
+    info("The extractor can extract them wherever they might be")
     Given("a document containing concepts in various places")
     // as a property of the top-level object
     // in a list
@@ -176,13 +156,12 @@ class ConceptExtractorTest
        |  }
        |}
        |}""".stripMargin
-    println(json)
     Then("all the concepts are returned")
     val concepts = ConceptExtractor(json)
     concepts.length shouldBe 8
   }
 
-  Scenario("Concepts with different labels but same id") {
+  Scenario("extract concepts with different labels but same id") {
     Given("a document with two concept objects")
     And("both concept objects have the same id, but different labels")
     val concept1 = SourceConcept(
@@ -203,18 +182,7 @@ class ConceptExtractorTest
     concepts.loneElement.label shouldBe "Isaac Newton"
   }
 
-  Scenario("A real example") {
-    // Choose an actual document from the API output and
-    // prove that it works.
-    val json = Source.fromResource("uk4kymkq.json").getLines().mkString("\n")
-    val concepts = ConceptExtractor(json)
-    // The more precisely defined tests here show what should be
-    // extracted.  This is just to demonstrate that it can successfully
-    // find the concepts in a real document from the catalogue api.
-    concepts.length shouldBe 6
-  }
-
-  Scenario("Source Concept with multiple identifiers") {
+  Scenario("extract a source Concept with multiple identifiers") {
     // If a source concept has multiple identifiers, then this
     // results in multiple concepts in the output.
     Given("a document with one concept object")
@@ -261,23 +229,24 @@ class ConceptExtractorTest
     concepts(1).identifier.identifierType.id shouldBe identifierType2
     concepts(1).identifier.value shouldBe identifier2
   }
-
+}
   Feature("Handling bad input") {
 
-    Scenario("A document with no Concepts") {
+    Scenario("a document with no Concepts") {
       Given("a json document with no concepts in it")
       val concepts = ConceptExtractor("""{"hello": "world"}""")
       Then("the concept list is empty")
       concepts shouldBe Nil
     }
 
-    Scenario("A document that is not JSON") {
+    Scenario("a document that is not JSON") {
       Given("an invalid document")
       val notJSON = "<hello>world</hello>"
       Then("an exception is raised")
       a[ParseException] should be thrownBy ConceptExtractor(notJSON)
     }
-    val t = Table(
+
+    val malformations = Table(
       ("malformation", "badJson"),
       (
         "an unknown identifier type",
@@ -347,8 +316,8 @@ class ConceptExtractorTest
                                    |}""".stripMargin
       )
     )
-    forAll(t) { (malformation, badJson) =>
-      Scenario(s"Malformed concept - $malformation") {
+    forAll(malformations) { (malformation, badJson) =>
+      Scenario(s"encountering a malformed concept - $malformation") {
         Given("a document with a valid concept and an invalid concept")
         val jsonString = s"""
         |{
@@ -370,10 +339,5 @@ class ConceptExtractorTest
         concepts.loneElement.canonicalId shouldBe "92345678"
       }
     }
-  }
-  Feature("Storing it (this is somewhere else)") {
-
-    // How do we handle upserts, when there may be alternative labels
-    // It may have to be fetch and update, or will it be
   }
 }
